Under certain conditions connections to the database would not be stopped in case of an error
ZongJi opens two connections to the MySQL database, one for the binlog events, another one for getting schema and to kill the binlog connection. Under certain conditions it can happen, that the control connection is not stopped. But as mysql-events tries to reconnect in the error handler it produces a new connection to the MySQL server every 4 seconds. In our case we quickly reached the connection limit and all other applications stopped working as their connection attempt got rejected.